# swagger.yaml
openapi: 3.0.3
info:
  title: API de Monitoreo de Calidad del Aire
  description: API REST para el sistema de monitoreo de calidad del aire con arquitectura hexagonal
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: soporte@monitoreo-aire.com
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
security:
  - bearerAuth: []
  - deviceToken: []
paths:
...
  /assignments:
    post:
      summary: Crear asignación
      description: Crea una nueva asignación de limpieza basada en una lectura
      tags:
        - Asignaciones
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id_lectura
                - id_usuario_asignado
                - descripcion
              properties:
                id_lectura:
                  type: integer
                  example: 1
                id_usuario_asignado:
                  type: integer
                  example: 2
                descripcion:
                  type: string
                  example: "Limpiar filtros de ventilación en Sala A - Alta concentración de PM2.5 detectada"
                prioridad:
                  type: string
                  enum: [baja, media, alta, urgente]
                  default: media
                  example: "alta"
      responses:
        201:
          description: Asignación creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asignacion'
        403:
          description: Permisos insuficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Obtener asignaciones
      description: Lista asignaciones con filtros opcionales
      tags:
        - Asignaciones
      security:
        - bearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, en_progreso, completada, cancelada]
          description: Filtrar por estado
          example: "pendiente"
        - name: user_id
          in: query
          schema:
            type: integer
          description: Filtrar por usuario asignado
          example: 2
      responses:
        200:
          description: Lista de asignaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asignacion'
  /assignments/{assignmentId}:
    get:
      summary: Obtener asignación específica
      description: Obtiene una asignación específica por su ID
      tags:
        - Asignaciones
      security:
        - bearerAuth: []
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        200:
          description: Información de la asignación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asignacion'
        404:
          description: Asignación no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Actualizar asignación
      description: Actualiza el estado de una asignación
      tags:
        - Asignaciones
      security:
        - bearerAuth: []
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - estado
              properties:
                estado:
                  type: string
                  enum: [pendiente, en_progreso, completada, cancelada]
                  example: "completada"
      responses:
        200:
          description: Asignación actualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Asignación actualizada a estado: completada"
        403:
          description: Permisos insuficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Asignación no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Asignacion:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_lectura:
          type: integer
          example: 1
        id_usuario_asignado:
          type: integer
          example: 2
        descripcion:
          type: string
          example: "Limpiar filtros de ventilación en Sala A - Alta concentración de PM2.5 detectada"
        estado:
          type: string
          enum: [pendiente, en_progreso, completada, cancelada]
          example: "pendiente"
        prioridad:
          type: string
          enum: [baja, media, alta, urgente]
          example: "alta"
        fecha_asignacion:
          type: string
          format: date-time
          example: "2025-01-15T14:35:00Z"
        fecha_completado:
          type: string
          format: date-time
          nullable: true
          example: null
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Mensaje de error descriptivo"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /login

    deviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Token único del dispositivo para autenticación de IoT